refactor: split server.py (3,775 lines) into server/ package with route modules
Breaks the monolithic server.py into focused modules under helix_context/server/:

  __init__.py         (42 lines)  - backward-compat re-exports
  app.py              (223 lines) - create_app() factory, lifespan, telemetry init
  helpers.py          (966 lines) - shared helpers, background tasks, proxy functions
  routes_context.py   (870 lines) - /context, /context/packet, /context/refresh-plan,
                                    /v1/chat/completions, /fingerprint
  routes_ingest.py    (148 lines) - /ingest, /consolidate
  routes_registry.py  (349 lines) - /sessions/*, /hitl/*, /context/expand, /session/manifest
  routes_admin.py     (884 lines) - /admin/*, /stats, /health, /debug/*, /bridge/*, /replicas

All 1933 tests pass with zero logic changes. Endpoint PRs no longer all diff
against one 3,775-line file.
